Moving 1.2 code to common, removing version-specific imports
