+++ package.json
@@ -65,8 +65,8 @@
   "dependencies": {
     "@emotion/react": "^11.8.1",
     "@emotion/styled": "^11.8.1",
-    "@finos/fdc3-2.0": "^2.0.0-beta.4",
-    "@finos/fdc3-1.2": "npm:@finos/fdc3@^1.2.0",
+    "fdc3-2.0": "npm:@finos/fdc3@^2.0.0-beta.4",
+    "fdc3-1.2": "npm:@finos/fdc3@^1.2.0",
     "@fontsource/roboto": "^4.5.3",
     "@mui/icons-material": "^5.10.15",
     "@mui/lab": "^5.0.0-alpha.109",
