--- packages/main/src/handlers/fdc3/1.2/topics.ts
@@ -1,22 +1,0 @@
-export const FDC3_1_2_TOPICS = {
-  // DROP_CONTEXT_LISTENER: 'FDC3_1.2:dropContextListener',
-  // ADD_CONTEXT_LISTENER: 'FDC3_1.2:addContextListener',
-  // DROP_INTENT_LISTENER: 'FDC3_1.2:dropIntentListener',
-  // ADD_INTENT_LISTENER: 'FDC3_1.2:addIntentListener',
-  // BROADCAST: 'FDC3_1.2:broadcast',
-  // OPEN: 'FDC3_1.2:open',
-  // CONTEXT: 'FDC3_1.2:context',
-  // GET_CURRENT_CONTEXT: 'FDC3_1.2:getCurrentContext',
-  // GET_OR_CREATE_CHANNEL: 'FDC3_1.2:getOrCreateChannel',
-  // GET_SYSTEM_CHANNELS: 'FDC3_1.2:getSystemChannels',
-  // LEAVE_CURRENT_CHANNEL: 'FDC3_1.2:leaveCurrentChannel',
-  // GET_CURRENT_CHANNEL: 'FDC3_1.2:getCurrentChannel',
-  // JOIN_CHANNEL: 'FDC3_1.2:joinChannel',
-  // INTENT: 'FDC3_1.2:intent',
-  // FIND_INTENT: 'FDC3_1.2:findIntent',
-  // FIND_INTENTS_BY_CONTEXT: 'FDC3_1.2:findIntentsByContext',
-  // RAISE_INTENT: 'FDC3_1.2:raiseIntent',
-  // RAISE_INTENT_FOR_CONTEXT: 'FDC3_1.2:raiseIntentForContext',
-  // RESOLVE_INTENT: 'FDC3_1.2:resolveIntent',
-  // INTENT_LISTENER_CALLED: 'FDC3_1_2:intentListenerCalled',
-};
