+++ packages/main/src/handlers/fdc3/lib/findIntent.ts
@@ -1,20 +1,21 @@
 import { getRuntime } from '/@/index';
-import { AppIntent, AppMetadata, ResolveError } from 'fdc3-1.2';
 import { DirectoryApp, DirectoryIntent } from '/@/directory/directory';
 import {
   FDC3Message,
   FindIntentData,
   FindIntentContextData,
+  SailAppIntent,
 } from '/@/types/FDC3Message';
+import { NoAppsFound } from '/@/types/FDC3Errors'; 
 
-function convertApp(a: DirectoryApp): AppMetadata {
-  return {
-    name: a.name ?? a.appId ?? '',
-    title: a.title,
-    description: a.description,
-    icons: a?.icons?.map((i) => i.src ?? '') ?? [],
-  };
-}
+// function convertApp(a: DirectoryApp): SailAppMetadata {
+//   return {
+//     name: a.name ?? a.appId ?? '',
+//     title: a.title,
+//     description: a.description,
+//     icons: a?.icons?.map((i) => i.src ?? '') ?? [],
+//   };
+// }
 
 export const findIntent = async (message: FDC3Message) => {
   const runtime = getRuntime();
@@ -29,15 +30,15 @@ export const findIntent = async (message: FDC3Message) => {
   );
 
   if (result.length == 0) {
-    throw new Error(ResolveError.NoAppsFound);
+    throw new Error(NoAppsFound);
   }
 
   const intentDisplayName =
     dir.retrieveAllIntentsByName(intent)[0]?.displayName ?? intent;
 
-  const r: AppIntent = {
+  const r: SailAppIntent = {
     intent: { name: intent, displayName: intentDisplayName },
-    apps: result.map(convertApp),
+    apps: result,
   };
 
   return r;
@@ -53,8 +54,8 @@ export const findIntentsByContext = async (message: FDC3Message) => {
       .getDirectory()
       .retrieveAllIntentsByContext(context.type);
 
-    const result: AppIntent[] = Object.keys(matchingIntents).map((k) => {
-      const apps = matchingIntents[k].map((o) => convertApp(o.app));
+    const result: SailAppIntent[] = Object.keys(matchingIntents).map((k) => {
+      const apps : DirectoryApp[] = matchingIntents[k].map((o) => o.app);
 
       return {
         intent: {
@@ -62,11 +63,11 @@ export const findIntentsByContext = async (message: FDC3Message) => {
           displayName: matchingIntents[k][0].displayName,
         },
         apps: apps,
-      } as AppIntent;
+      } as SailAppIntent;
     });
     if (result.length > 0) {
       return result;
     }
   }
-  throw new Error(ResolveError.NoAppsFound);
+  throw new Error(NoAppsFound);
 };
