+++ packages/main/src/handlers/fdc3/lib/open.ts
@@ -5,7 +5,8 @@ import {
   DirectoryAppLaunchDetailsWeb,
 } from '/@/directory/directory';
 import { getSailManifest } from '/@/directory/directory';
-import { Context } from '@finos/fdc3';
+import { Context, FDC3Message, OpenData } from '/@/types/FDC3Message';
+import { AppNotFound } from '/@/types/FDC3Errors';
 
 export const openApp = async (
   app: DirectoryApp,
@@ -37,4 +38,28 @@ export const openApp = async (
     newView.setPendingContext(context, source);
   }
   return;
+};
+
+export const open = async (message: FDC3Message) => {
+  const runtime = getRuntime();
+  const data: OpenData = message.data as OpenData;
+  const targetIdentifier = data.target;
+
+  let allResults : DirectoryApp[] = [];
+
+  if (targetIdentifier.name) {
+    allResults = runtime.getDirectory().retrieveByName(targetIdentifier.name);
+  } else if (targetIdentifier.appId) {
+    allResults = runtime.getDirectory().retrieveByAppId(targetIdentifier.appId);
+  } else if (targetIdentifier.key) {
+    allResults = runtime.getDirectory().retrieveByAppId(targetIdentifier.key);
+  }
+
+  if (allResults.length > 0) {
+    // this deals with the case of multiple apps having the same name, appId etc.
+    // TODO: if there's more than one, should this be an error?
+    const directoryEntry: DirectoryApp = allResults[0];
+    return await openApp(directoryEntry, message.source, data.context);
+  }
+  throw new Error(AppNotFound);
 };
