+++ packages/main/src/view.ts
@@ -9,7 +9,6 @@ import { ViewConfig } from './types/ViewConfig';
 import { getRuntime } from './index';
 import { BrowserView } from 'electron';
 import { DirectoryApp } from './directory/directory';
-import { Context } from '@finos/fdc3';
 import { Rectangle } from 'electron/main';
 import { Workspace } from './workspace';
 import { FDC3Listener } from './types/FDC3Listener';
