+++ packages/main/tsconfig.json
